+++ pakyow-core/CHANGELOG.md
@@ -5,6 +5,7 @@
   * [fixed] App connection path is relative to to the app mount path
   * [fixed] Methods defined an an app block are now correctly defined
   * [fixed] Issue with `Pakyow::Error` not detecting gems in rvm
+  * [fixed] Query string missing from normalized uris
 
 # 1.0.1